refactor: change menu classes structure by BOUNDED CONTEXT
